+++ lib/command/convert.rb
@@ -172,6 +172,9 @@ module Command
     def convert_novels(argv)
       tagname_to_ids(argv)
       argv.each.with_index(1) do |target, i|
+        @target = target
+        @novel_data = nil
+
         Helper.print_horizontal_rule if i > 1
         if @basename
           @basename << " (#{i})" if argv.length > 1